Created base project setup: README, build-chain, and initial files.
In README.md:
Created initial readme file for the project's basic info. Now has most of the basic goals and strategies for the project, but still needs more work. It does have all of the notes from the paper copy of the initial specification, with a few minor edits.

In Gulpfile.js:
Created as bare-bones base to start a build-chain off of for better integration with this projects tools. But it still has many MAJOR issues, such as the "gulp-prettier" plug-in causing CATASTROPHIC data-loss! The root cause of the problem has not yet been fully tracked down, but it is known that having a Gulp Watch setting of less than 200-milliseconds is EXTREMELY problematic! So until it's known what exactly cause the problem, Gulp should NOT be used with its current settings (especially with tools that write to files), if at all!
More detail is in the "warning-gulp-misuse" directory.

In Prettier RC Config:
Most Prettier preferences have been set except for the curly-bracket and parenthesises spacing settings. It works quite well, but its best use is certainly in IDE/editor integration.

In ESLint RC Config:
While most ESLint rules havn't been set yet, it is already clear it's almost impractically useless when not used as an IDE/editor integration. The rules for prettier and TypeScript integration are already working well.

In TSConfig.json:
Setting have been working quite well for a very long time. But adjustments to the output JavaScript of the TSC could be changed to better allow to JavaScript-only devs to make meaninful contributions to the project, likely ECMAScript output verion, config for using ts.d files, formatting of JS output, etc.

In .VSCode:
Added JSON SchemaStore entries.

In Source TypeScript:
Project still in progress. Basic CLI is partially done.