Improved build-chain and other updates.
In /src/:
	Formatted, linted, and spell-checked. Needs to be updated and tested
with newer versions of dependencies. The `OpsPipelineQueue` code should
be moved to a separate project, and either made a dependency or a git
sub-module if possible. `union-helper-types.ts` and other type helpers
should be moved into other projects or replaced with tools that handle
runtime parsing and validation more robustly.

In /esbuild.mjs:
	ESBuild is now used for transpiling TS and bundling. A single script
is much more ergonomic to test and run.

In /.husky/:
	Added git pre-commit hook to check linting and formatting, and run
test suite.

In /__tests__/:
	Added tests for NPM CLI API compatibility. Formatted, linted, and
spell-checked other tests. Also updated some typings for newer TS
versions and added some better practices.

In /dist/:
	Removed old TSC outDir since ESBuild is now being used.

In /bin/:
	Is now used for ESBuild outFile. Removed TS configuration.

In TypeScript Config:
	Removed build-chain configuration not needed after transition
to ESBuild.

In /npm-backup.sh:
	Shell script was never really worked on, and it is less portable that
NodeJS. All work on a shell script in now cancled and removed from any
future plans.

In /:
	Cleaned up NPM config and updated packages. Updated formatting and
linting config. Spell-checked most files, and added cspell config.